+++ pushtemplates/src/main/java/com/clevertap/pushtemplates/PushTemplateMessagingService.java
@@ -3,6 +3,7 @@ package com.clevertap.pushtemplates;
 import android.content.Context;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import com.clevertap.android.sdk.CleverTapAPI;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
@@ -14,7 +15,7 @@ public class PushTemplateMessagingService extends FirebaseMessagingService {
     Context context;
 
     @Override
-    public void onMessageReceived(RemoteMessage remoteMessage) {
+    public void onMessageReceived(@NonNull RemoteMessage remoteMessage) {
         try {
             PTLog.debug("Inside Push Templates");
             context = getApplicationContext();
@@ -37,5 +38,10 @@ public class PushTemplateMessagingService extends FirebaseMessagingService {
         } catch (Throwable throwable) {
             PTLog.verbose("Error parsing FCM payload", throwable);
         }
+    }
+
+    @Override
+    public void onNewToken(@NonNull final String s) {
+
     }
 }